+++ packages/terra-core-docs/src/terra-dev-site/test/action-header/PreviousEnabledNextDisabledActionHeader.test.jsx
@@ -4,8 +4,9 @@ import ActionHeader from 'terra-action-header/lib/ActionHeader';
 
 const ActionHeaderExample = () => (
   <ActionHeader
-    title="Previous Enabled Next Disabled Action Header"
+    text="Previous Enabled Next Disabled Action Header"
     onPrevious={() => alert('You clicked previous!')}
+    level={1}
   />
 );