+++ packages/terra-action-header/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+  * Breaking Changes
+    * Renamed `title` prop as `text` to avoid confusion with HTML attribute `title`.
+    * Removed default `h1` heading level.
+    
 ## 2.74.3 - (July 5, 2022)
 
 * Changed